skin-recommender-api/requirements.txt
Flask
tensorflow
Pillow
numpy
Flask-Cors
Flask-Bcrypt
Flask-JWT-Extended